Move expression context into its own file
